refactor: 랜덤 매칭 로직 유틸 및 커스텀 훅으로 분리 (#113)
