+++ src/app/team-maker/random-matching/page.tsx
@@ -1,94 +1,23 @@
 "use client";
 
-import { row1Names, row2Names, row3Names } from "@/constants/team-maker";
-import { useEffect, useState } from "react";
-
-const getRandomRow = (sourceNames) => {
-  const row = [];
-  for (let i = 0; i < 20; i++) {
-    row.push(sourceNames[Math.floor(Math.random() * sourceNames.length)]);
-  }
-  return row;
-};
+import MatchiingSuccessModal from "@/components/TeamMaker/Modal/MatchingSuccessModal";
+import RandomMatching from "@/components/TeamMaker/RandomMatching/RandomMatching";
+import useToggle from "@/hooks/useToggle";
+import mockData from "@/mocks/teamDTOList.json";
 
 export default function RandomMatchingPage() {
-  const [rows, setRows] = useState([
-    getRandomRow(row1Names),
-    getRandomRow(row2Names),
-    getRandomRow(row3Names),
-  ]);
-  const [offsets, setOffsets] = useState([0, 0, 0]);
-  const [stopped, setStopped] = useState(false);
+  const { isOpen, toggle } = useToggle();
 
-  useEffect(() => {
-    if (stopped) return;
-
-    const interval = setInterval(() => {
-      setOffsets(([o1, o2, o3]) => [o1 + 1, o2 - 1, o3 + 1]);
-    }, 100);
-
-    const stopTimeout = setTimeout(() => {
-      clearInterval(interval);
-      setStopped(true);
-    }, 3000);
-
-    return () => {
-      clearInterval(interval);
-      clearTimeout(stopTimeout);
-    };
-  }, [stopped]);
-
-  const getVisibleRow = (row, offset) => {
-    const visible = [];
-    for (let i = -3; i <= 3; i++) {
-      const index = (((offset + i) % row.length) + row.length) % row.length;
-      visible.push(row[index]);
-    }
-    return visible;
-  };
-
-  const visibleRows = rows.map((row, i) => getVisibleRow(row, offsets[i]));
-  // const centerTeam = visibleRows.map((row) => row[2]);
+  const { teamDtoList } = mockData;
 
   return (
-    <div className="flex flex-col items-center justify-center pt-[120px]">
-      <div className="flex flex-col items-center mb-[91px]">
-        <span className="w-[121px] h-[53px] text-[24px] font-bold px-4 py-3 mb-3 rounded-[4px] bg-primary">
-          팀 메이커
-        </span>
-        <h1 className="h-[60px] text-[36px] mb-11 font-bold leading-[60px]">
-          랜덤 매칭
-        </h1>
-        <p className="text-center text-[48px] font-semibold leading-[60px]">
-          팀메이커로 랜덤 매칭하고 <br />
-          나와 함께
-          <span className="text-primary"> 협업할 팀원 </span>
-          찾아가세요!
-        </p>
-      </div>
-
-      <div className="mb-[243px]">
-        {visibleRows.map((row, rowIndex) => (
-          <div key={rowIndex} className="flex justify-center min-h-[106px]">
-            {row.map((name, idx) => (
-              <div
-                key={`${rowIndex}-${idx}`}
-                className={`px-[47px] py-[32px] text-[32px] font-semibold text-white leading-10 text-center min-w-[178px] ${
-                  idx === 3 ? "bg-orange-500" : "bg-gray5"
-                }`}
-              >
-                {name}
-              </div>
-            ))}
-          </div>
-        ))}
-      </div>
-
-      {/* {stopped && (
-        <div className="mt-10 text-lg">
-          매칭된 팀원: <strong>{centerTeam.join(", ")}</strong>
-        </div>
-      )} */}
-    </div>
+    <>
+      <RandomMatching toggle={toggle} />
+      <MatchiingSuccessModal
+        isOpen={isOpen}
+        toggle={toggle}
+        teamData={teamDtoList}
+      />
+    </>
   );
 }
